MCR-2464 Move resolver into separate public class that is indididually enabled the configuration
