+++ mycore-base/src/main/resources/config/mycore.properties
@@ -56,6 +56,7 @@
 # Sortable flag for Classification with ID {MCRObjectID}
 # MCR.URIResolver.Classification.Sort.{MCRObjectID}=true
 
+  MCR.URIResolver.ModuleResolver.properties=org.mycore.common.xml.MCRPropertiesResolver
   MCR.URIResolver.ModuleResolver.userobjectrights=org.mycore.common.xml.MCRUserAndObjectRightsURIResolver
 
 ##############################################################################
